v 0.1

+++ Order/src/main/java/ru/t1/Order/controller/UserController.java
@@ -2,10 +2,11 @@ package ru.t1.Order.controller;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
+import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 import ru.t1.Order.dto.OrderDTO;
 import ru.t1.Order.dto.UserDTO;
-import ru.t1.Order.exception.UserAlreadyExistException;
 import ru.t1.Order.exception.UserNotFoundException;
 import ru.t1.Order.service.UserService;
 
@@ -20,49 +21,80 @@ public class UserController {
         this.userService = userService;
     }
 
-    @PostMapping
-    public ResponseEntity<UserDTO> createUser(@RequestBody UserDTO user) {
-        try {
-            UserDTO createdUser = userService.createUser(user);
-            return ResponseEntity.status(HttpStatus.CREATED).body(createdUser);
-        } catch (UserAlreadyExistException e) {
-            return ResponseEntity.status(HttpStatus.CONFLICT).build();
-        }
-    }
 
     @PutMapping
-    public ResponseEntity<UserDTO> updateUser(@RequestBody UserDTO user) {
+    @PreAuthorize("hasRole('USER')")
+    public ResponseEntity<UserDTO> updateUser(@RequestBody UserDTO user, Authentication authentication) {
         try {
+            String authName = authentication.getName();
             UserDTO updatedUser = userService.updateUser(user);
-            return ResponseEntity.ok(updatedUser);
+
+            if (authName.equals(updatedUser.getName())) {
+                return ResponseEntity.ok(updatedUser);
+            } else {
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            }
         } catch (UserNotFoundException e) {
             return ResponseEntity.notFound().build();
         }
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<UserDTO> getUser(@PathVariable int id) {
+    @PreAuthorize("hasRole('USER') or hasRole('ADMIN')")
+    public ResponseEntity<UserDTO> getUser(@PathVariable int id, Authentication authentication) {
         try {
+            String authName = authentication.getName();
             UserDTO user = userService.getUser(id);
-            return ResponseEntity.ok(user);
+
+            if (authentication.getAuthorities().stream().anyMatch(
+                    authority -> authority.getAuthority().equals("ROLE_ADMIN")) ||
+                    user.getName().equals(authName)) {
+                return ResponseEntity.ok(user);
+            } else {
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            }
         } catch (UserNotFoundException e) {
             return ResponseEntity.notFound().build();
         }
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteUser(@PathVariable int id) {
+    @PreAuthorize("hasRole('USER') or hasRole('ADMIN')")
+    public ResponseEntity<Void> deleteUser(@PathVariable int id, Authentication authentication) {
         try {
-            userService.deleteUser(id);
-            return ResponseEntity.ok().build();
+            String authName = authentication.getName();
+            UserDTO user = userService.getUser(id);
+
+            if (authentication.getAuthorities().stream().anyMatch(
+                    authority -> authority.getAuthority().equals("ROLE_ADMIN"))
+                    || user.getName().equals(authName)) {
+                userService.deleteUser(id);
+                return ResponseEntity.ok().build();
+            } else {
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            }
         } catch (UserNotFoundException e) {
             return ResponseEntity.notFound().build();
         }
     }
 
     @GetMapping("/orders/{userId}")
-    public ResponseEntity<List<OrderDTO>> getOrdersByUserId(@PathVariable int userId) {
-        List<OrderDTO> orders = userService.getOrders(userId);
-        return ResponseEntity.ok(orders);
+    @PreAuthorize("hasRole('USER') or hasRole('ADMIN')")
+    public ResponseEntity<List<OrderDTO>> getOrdersByUserId(@PathVariable int userId, Authentication authentication) {
+        try {
+            String authName = authentication.getName();
+            UserDTO user = userService.getUser(userId);
+
+            if (authentication.getAuthorities().stream().anyMatch(
+                    authority -> authority.getAuthority().equals("ROLE_ADMIN")) ||
+                    user.getName().equals(authName)) {
+                List<OrderDTO> orders = userService.getOrders(userId);
+                return ResponseEntity.ok(orders);
+            } else {
+                return ResponseEntity.status(HttpStatus.FORBIDDEN).build();
+            }
+        } catch (UserNotFoundException e) {
+            return ResponseEntity.notFound().build();
+        }
     }
 }
